requirements.txt
httpx
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pandas
pydantic
fastapi
uvicorn
icecream
gpxpy
geopandas
streamlit
streamlit-folium
streamlit-authenticator
streamlit-extras
streamlit-calendar
streamlit-card
streamlit-local-storage
PyGithub
plotly
pyotp
qrcode
celery
redis
flower
crontab
psycopg2
sqlalchemy
backtesting